requirements.txt
dj-database-url==0.4.1
Django==1.9
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
djangorestframework==3.6.2
django-filter==1.0.2
Markdown==2.6.8
